+++ sentry-android-core/src/main/java/io/sentry/android/core/AndroidProfiler.java
@@ -4,14 +4,11 @@ import android.annotation.SuppressLint;
 import android.os.Debug;
 import android.os.Process;
 import android.os.SystemClock;
-import io.sentry.CpuCollectionData;
 import io.sentry.DateUtils;
 import io.sentry.ILogger;
 import io.sentry.ISentryExecutorService;
 import io.sentry.ISentryLifecycleToken;
-import io.sentry.MemoryCollectionData;
 import io.sentry.PerformanceCollectionData;
-import io.sentry.SentryDate;
 import io.sentry.SentryLevel;
 import io.sentry.SentryNanotimeDate;
 import io.sentry.SentryUUID;
@@ -156,7 +153,7 @@ public class AndroidProfiler {
                   // profileStartNanos is calculated through SystemClock.elapsedRealtimeNanos(),
                   // but frameEndNanos uses System.nanotime(), so we convert it to get the timestamp
                   // relative to profileStartNanos
-                  final SentryDate timestamp = new SentryNanotimeDate();
+                  final long timestampNanos = new SentryNanotimeDate().nanoTimestamp();
                   final long frameTimestampRelativeNanos =
                       frameEndNanos
                           - System.nanoTime()
@@ -171,17 +168,17 @@ public class AndroidProfiler {
                   if (isFrozen) {
                     frozenFrameRenderMeasurements.addLast(
                         new ProfileMeasurementValue(
-                            frameTimestampRelativeNanos, durationNanos, timestamp));
+                            frameTimestampRelativeNanos, durationNanos, timestampNanos));
                   } else if (isSlow) {
                     slowFrameRenderMeasurements.addLast(
                         new ProfileMeasurementValue(
-                            frameTimestampRelativeNanos, durationNanos, timestamp));
+                            frameTimestampRelativeNanos, durationNanos, timestampNanos));
                   }
                   if (refreshRate != lastRefreshRate) {
                     lastRefreshRate = refreshRate;
                     screenFrameRateMeasurements.addLast(
                         new ProfileMeasurementValue(
-                            frameTimestampRelativeNanos, refreshRate, timestamp));
+                            frameTimestampRelativeNanos, refreshRate, timestampNanos));
                   }
                 }
               });
@@ -318,32 +315,28 @@ public class AndroidProfiler {
           new ArrayDeque<>(performanceCollectionData.size());
 
       synchronized (performanceCollectionData) {
-        for (PerformanceCollectionData performanceData : performanceCollectionData) {
-          CpuCollectionData cpuData = performanceData.getCpuData();
-          MemoryCollectionData memoryData = performanceData.getMemoryData();
-          if (cpuData != null) {
+        for (final @NotNull PerformanceCollectionData data : performanceCollectionData) {
+          final long nanoTimestamp = data.getNanoTimestamp();
+          final long relativeStartNs = nanoTimestamp + timestampDiff;
+          final @Nullable Double cpuUsagePercentage = data.getCpuUsagePercentage();
+          final @Nullable Long usedHeapMemory = data.getUsedHeapMemory();
+          final @Nullable Long usedNativeMemory = data.getUsedNativeMemory();
+
+          if (cpuUsagePercentage != null) {
             cpuUsageMeasurements.add(
-                new ProfileMeasurementValue(
-                    cpuData.getTimestamp().nanoTimestamp() + timestampDiff,
-                    cpuData.getCpuUsagePercentage(),
-                    cpuData.getTimestamp()));
+                new ProfileMeasurementValue(relativeStartNs, cpuUsagePercentage, nanoTimestamp));
           }
-          if (memoryData != null && memoryData.getUsedHeapMemory() > -1) {
+          if (usedHeapMemory != null) {
             memoryUsageMeasurements.add(
-                new ProfileMeasurementValue(
-                    memoryData.getTimestamp().nanoTimestamp() + timestampDiff,
-                    memoryData.getUsedHeapMemory(),
-                    memoryData.getTimestamp()));
+                new ProfileMeasurementValue(relativeStartNs, usedHeapMemory, nanoTimestamp));
           }
-          if (memoryData != null && memoryData.getUsedNativeMemory() > -1) {
+          if (usedNativeMemory != null) {
             nativeMemoryUsageMeasurements.add(
-                new ProfileMeasurementValue(
-                    memoryData.getTimestamp().nanoTimestamp() + timestampDiff,
-                    memoryData.getUsedNativeMemory(),
-                    memoryData.getTimestamp()));
+                new ProfileMeasurementValue(relativeStartNs, usedNativeMemory, nanoTimestamp));
           }
         }
       }
+
       if (!cpuUsageMeasurements.isEmpty()) {
         measurementsMap.put(
             ProfileMeasurement.ID_CPU_USAGE,