+++ sentry/src/test/java/io/sentry/JavaMemoryCollectorTest.kt
@@ -2,8 +2,7 @@ package io.sentry
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import kotlin.test.assertNotEquals
-import kotlin.test.assertNotNull
+import kotlin.test.assertNull
 
 class JavaMemoryCollectorTest {
   private val fixture = Fixture()
@@ -15,13 +14,11 @@ class JavaMemoryCollectorTest {
 
   @Test
   fun `when collect, only heap memory is collected`() {
-    val data = PerformanceCollectionData()
+    val data = PerformanceCollectionData(10)
     val usedMemory = fixture.runtime.totalMemory() - fixture.runtime.freeMemory()
     fixture.collector.collect(data)
-    val memoryData = data.memoryData
-    assertNotNull(memoryData)
-    assertEquals(-1, memoryData.usedNativeMemory)
-    assertEquals(usedMemory, memoryData.usedHeapMemory)
-    assertNotEquals(0, memoryData.timestamp.nanoTimestamp())
+    assertNull(data.usedNativeMemory)
+    assertEquals(usedMemory, data.usedHeapMemory)
+    assertEquals(10, data.nanoTimestamp)
   }
 }